Add type annotations to `core/builtin_funcs.py` and `core/builtin_classes/base_classes.py`

+++ core/builtin_classes/requests_object.py
@@ -13,33 +13,50 @@ import urllib.parse
 class RequestsObject(BuiltInObject):
     @operator("__constructor__")
     @check([], [])
-    def constructor(self):
-        return RTResult().success(None)
+    def constructor(self) -> RTResult[Value]:
+        return RTResult[Value]().success(Null.null())
 
-    @args(["url", "headers"], [String, HashMap({})])
+    @args(["url", "headers"], [None, HashMap({})])
     @method
-    def get(ctx):
-        res = RTResult()
+    def get(self, ctx: Context) -> RTResult[Value]:
+        res = RTResult[Value]()
         url = ctx.symbol_table.get("url")
+        assert url is not None
+        if not isinstance(url, String):
+            return res.failure(RTError(url.pos_start, url.pos_end, "Expected String", ctx))
         headers = ctx.symbol_table.get("headers")
+        assert headers is not None
+        if not isinstance(headers, HashMap):
+            return res.failure(RTError(headers.pos_start, headers.pos_end, "Expected HashMap", ctx))
         try:
-            req = urllib.request.Request(url.value, headers=deradonify(headers))
+            req = urllib.request.Request(url.value, headers=deradonify(headers))  # type: ignore
             with urllib.request.urlopen(req) as response:
                 response_data = response.read().decode("utf-8")
             return res.success(radonify(response_data, url.pos_start, url.pos_end, url.context))
         except Exception as e:
             return res.failure(RTError(url.pos_start, url.pos_end, f"Error sending GET request: {str(e)}", ctx))
 
-    @args(["url", "data", "headers"], [String, HashMap({}), HashMap({})])
+    @args(["url", "data", "headers"], [None, HashMap({}), HashMap({})])
     @method
-    def post(ctx):
-        res = RTResult()
+    def post(self, ctx: Context) -> RTResult[Value]:
+        res = RTResult[Value]()
         url = ctx.symbol_table.get("url")
+        assert url is not None
+        if not isinstance(url, String):
+            return res.failure(RTError(url.pos_start, url.pos_end, "Expected String", ctx))
         data = ctx.symbol_table.get("data")
+        assert data is not None
         headers = ctx.symbol_table.get("headers")
+        assert headers is not None
+        if not isinstance(data, HashMap):
+            return res.failure(RTError(data.pos_start, data.pos_end, "Expected HashMap", ctx))
+        if not isinstance(headers, HashMap):
+            return res.failure(RTError(headers.pos_start, headers.pos_end, "Expected HashMap", ctx))
         try:
             req = urllib.request.Request(
-                url.value, data=json.dumps(deradonify(data)).encode("utf-8"), headers=deradonify(headers)
+                url.value,
+                data=json.dumps(deradonify(data)).encode("utf-8"),
+                headers=deradonify(headers),  # type: ignore
             )
             with urllib.request.urlopen(req) as response:
                 response_data = response.read().decode("utf-8")
@@ -47,16 +64,26 @@ class RequestsObject(BuiltInObject):
         except Exception as e:
             return res.failure(RTError(url.pos_start, url.pos_end, f"Error sending POST request: {str(e)}", ctx))
 
-    @args(["url", "data", "headers"], [String, HashMap({}), HashMap({})])
+    @args(["url", "data", "headers"], [None, HashMap({}), HashMap({})])
     @method
-    def put(ctx):
-        res = RTResult()
+    def put(self, ctx: Context) -> RTResult[Value]:
+        res = RTResult[Value]()
         url = ctx.symbol_table.get("url")
+        assert url is not None
+        if not isinstance(url, String):
+            return res.failure(RTError(url.pos_start, url.pos_end, "Expected String", ctx))
         data = ctx.symbol_table.get("data")
+        assert data is not None
         headers = ctx.symbol_table.get("headers")
+        assert headers is not None
+        if not isinstance(headers, HashMap):
+            return res.failure(RTError(headers.pos_start, headers.pos_end, "Expected HashMap", ctx))
         try:
             req = urllib.request.Request(
-                url.value, data=json.dumps(deradonify(data)).encode("utf-8"), headers=deradonify(headers), method="PUT"
+                url.value,
+                data=json.dumps(deradonify(data)).encode("utf-8"),
+                headers=deradonify(headers),  # type: ignore
+                method="PUT",
             )
             with urllib.request.urlopen(req) as response:
                 response_data = response.read().decode("utf-8")
@@ -64,32 +91,45 @@ class RequestsObject(BuiltInObject):
         except Exception as e:
             return res.failure(RTError(url.pos_start, url.pos_end, f"Error sending PUT request: {str(e)}", ctx))
 
-    @args(["url", "headers"], [String, HashMap({})])
+    @args(["url", "headers"], [None, HashMap({})])
     @method
-    def delete(ctx):
-        res = RTResult()
+    def delete(self, ctx: Context) -> RTResult[Value]:
+        res = RTResult[Value]()
         url = ctx.symbol_table.get("url")
+        assert url is not None
+        if not isinstance(url, String):
+            return res.failure(RTError(url.pos_start, url.pos_end, "Expected String", ctx))
         headers = ctx.symbol_table.get("headers")
+        assert headers is not None
+        if not isinstance(headers, HashMap):
+            return res.failure(RTError(headers.pos_start, headers.pos_end, "Expected HashMap", ctx))
         try:
-            req = urllib.request.Request(url.value, headers=deradonify(headers), method="DELETE")
+            req = urllib.request.Request(url.value, headers=deradonify(headers), method="DELETE")  # type: ignore
             with urllib.request.urlopen(req) as response:
                 response_data = response.read().decode("utf-8")
             return res.success(radonify(response_data, url.pos_start, url.pos_end, url.context))
         except Exception as e:
             return res.failure(RTError(url.pos_start, url.pos_end, f"Error sending DELETE request: {str(e)}", ctx))
 
-    @args(["url", "data", "headers"], [String, HashMap({}), HashMap({})])
+    @args(["url", "data", "headers"], [None, HashMap({}), HashMap({})])
     @method
-    def patch(ctx):
-        res = RTResult()
+    def patch(self, ctx: Context) -> RTResult[Value]:
+        res = RTResult[Value]()
         url = ctx.symbol_table.get("url")
+        assert url is not None
+        if not isinstance(url, String):
+            return res.failure(RTError(url.pos_start, url.pos_end, "Expected String", ctx))
         data = ctx.symbol_table.get("data")
+        assert data is not None
         headers = ctx.symbol_table.get("headers")
+        assert headers is not None
+        if not isinstance(headers, HashMap):
+            return res.failure(RTError(headers.pos_start, headers.pos_end, "Expected HashMap", ctx))
         try:
             req = urllib.request.Request(
                 url.value,
                 data=json.dumps(deradonify(data)).encode("utf-8"),
-                headers=deradonify(headers),
+                headers=deradonify(headers),  # type: ignore
                 method="PATCH",
             )
             with urllib.request.urlopen(req) as response:
